[NodeTypeResolver] split test per type
