+++ packages/NodeTypeResolver/tests/PerNodeTypeResolver/Source/SomeClass.php.inc
